Add basic plot function to cf_xarray
Not sure if I want to put logic for this on the individual
CFDiscreteGeometry classes, since I believe it would be
essentially the same on all of them - you always need to
convert to point array before plotting. Also need to consider
resampling to a raster vs. scatter-plotting, but I think this
functionality could only be on the grid-aware accessor.