R Remove beams_vars and ts_dtype from new readers
These are not used at all by the new readers at the moment.
Reader classes should handle expansion of variables along a dimension
in the pre- or post-processing functions if needed. As we are now
careful to pass along encoding information from start to finish when
reading and stacking, we also have no need to pre-define desired dtypes
for output. If these should be changed along the way, this can also be
done in pre- or post-processing.